+++ frontend/src/components/Banner.js
@@ -3,35 +3,17 @@
 
 import React from "react";
 import { useMediaQuery } from "react-responsive";
-
-import HeaderTypography from "./HeaderTypography";
-import ParagraphTypography from "./ParagraphTypography";
-import { Link } from "react-router-dom";
-import RoundedEdgesButton from "./RoundedEdgesButton";
-
-import { textColor, activeTextColor } from "../assets/jss/colorPalette";
 import { landingPageBannerImg } from "../assets/jss/imagePalette";
-
-/**
- * Displays a welcoming background image with a button.
- * @param user user object of the currently logged in account
- * @returns {JSX.Element}
- */
-export default function WelcomeBanner({ user }) {
-	const welcomeTitle = "Welcome to NFT The World!";
-	const welcomeSubtitle = <>
-		Breathtaking NFT shopping lounge.<br/>
-		Own your virtual map area from Nuremburg and Riga!
-	</>;
-	return <Banner user={user} title={welcomeTitle} subtitle={ welcomeSubtitle }/>;
-}
+import { BannerTitle } from "./BannerTitle";
+import { BannerSubTitle } from "./BannerSubTitle";
+import { BannerButton } from "./BannerButton";
 
 /**
  * Displays background banner with headline "title", text "subtitle" and login-dependent button.
  * @param user user object of the currently logged in account
  * @returns {JSX.Element}
  */
-export function Banner({ subtitle, title, user }) {
+export default function Banner({ subtitle, title, user }) {
 	const render=() => (
 		<div style={containerStyle}>
 			<div style={backgroundImageStyle}>
@@ -53,48 +35,4 @@ export function Banner({ subtitle, title, user }) {
 	const backgroundImageStyle = ({ padding: "44px", borderRadius: "15px", backgroundImage: `url(${landingPageBannerImg})`, backgroundSize: "cover", backgroundPosition: "center", alignItems: "center", justifyContent: "center", textAlign: "center", });
 	const bannerItemStyle = ({ width: "100%", display: "flex", alignItems: "center", justifyContent: "center" });
 	return render();
-}
-
-const bannerStyle = ({ color: textColor });
-
-function BannerTitle({ isLarge, children }) {
-	const smallStyle = { ...bannerStyle, fontSize: 30, fontWeight: 550, bottomMargin: "5px" };
-	const largeStyle = { ...bannerStyle, fontSize: 37, fontWeight: 700 };
-	const style = isLarge? largeStyle : smallStyle;
-
-	return <HeaderTypography style={style}>{ children }</HeaderTypography>;
-}
-
-function BannerSubTitle({ isLarge, children }) {
-	const smallStyle = { ...bannerStyle, fontSize: 19, fontWeight: 400 };
-	const largeStyle = { ...bannerStyle, fontSize: 22, fontWeight: 500 };
-	const style = isLarge? largeStyle : smallStyle;
-
-	return <ParagraphTypography style={style}>{ children }</ParagraphTypography>;
-}
-
-// according to mockup, the Banner's Button is slightly larger than the Header's
-function BannerButton({ isLoggedIn }) {
-	const buttonStyle = { height: "54px", width: "192px", fontWeight: 700, fontSize: "17px", };
-
-	const BannerButtonComponent =  isLoggedIn? ProfileButton : SignupButton;
-	return <BannerButtonComponent style={buttonStyle}/>;
-}
-
-function SignupButton({ style }) {
-	const signupButtonStyle = { backgroundColor: activeTextColor, ...style };
-	return (
-		<RoundedEdgesButton style={signupButtonStyle} component={Link} to="/signup">
-			Sign Up
-		</RoundedEdgesButton>
-	);
-}
-
-function ProfileButton({ style }) {
-	const profileButtonStyle = { backgroundColor: "black", color: textColor, ...style };
-	return (
-		<RoundedEdgesButton color="inherit" style={profileButtonStyle} component={Link} to="/user/profile">
-			Profile
-		</RoundedEdgesButton>
-	);
 }
